+++ src/buildtool/main/main.cpp
@@ -787,7 +787,8 @@ void DumpTrees(std::string const& file_path, AnalysisResult const& result) {
 }
 
 void DumpTargets(std::string const& file_path,
-                 std::vector<Target::ConfiguredTarget> const& target_ids) {
+                 std::vector<Target::ConfiguredTarget> const& target_ids,
+                 std::string const& target_qualifier = "") {
     auto repo_map = nlohmann::json::object();
     auto conf_list =
         [&repo_map](Base::EntityName const& ref) -> nlohmann::json& {
@@ -813,12 +814,14 @@ void DumpTargets(std::string const& file_path,
         });
     auto const dump_string = IndentListsOnlyUntilDepth(repo_map, 2);
     if (file_path == "-") {
-        Logger::Log(LogLevel::Info, "List of analysed targets:");
+        Logger::Log(
+            LogLevel::Info, "List of analysed {}targets:", target_qualifier);
         std::cout << dump_string << std::endl;
     }
     else {
         Logger::Log(LogLevel::Info,
-                    "Dumping list of analysed targets to file '{}'.",
+                    "Dumping list of analysed {}targets to file '{}'.",
+                    target_qualifier,
                     file_path);
         std::ofstream os(file_path);
         os << dump_string << std::endl;
@@ -940,6 +943,10 @@ void DumpNodes(std::string const& file_path, AnalysisResult const& result) {
     }
     if (clargs.dump_targets) {
         DumpTargets(*clargs.dump_targets, result_map.ConfiguredTargets());
+    }
+    if (clargs.dump_export_targets) {
+        DumpTargets(
+            *clargs.dump_export_targets, result_map.ExportTargets(), "export ");
     }
     if (clargs.dump_targets_graph) {
         auto graph = result_map.ConfiguredTargetsGraph().dump(2);